12-27 로그인/게시판

+++ Solo/src/main/java/com/proj/model/User.java
@@ -1,6 +1,6 @@
 package com.proj.model;
 
-import jakarta.persistence.Entity;
+import jakarta.persistence.Entity; 
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -15,7 +15,7 @@ import lombok.Setter;
 public class User {
 		@Id
 	    @GeneratedValue(strategy = GenerationType.IDENTITY)
-	    private int id;
+	    private Long id;
 		
 		private String username;
